+++ optuna/storages/__init__.py
@@ -1,8 +1,8 @@
 from typing import Union
 
-from optuna._callbacks import RetryFailedTrialCallback
 from optuna.storages._base import BaseStorage
 from optuna.storages._cached_storage import _CachedStorage
+from optuna.storages._callbacks import RetryFailedTrialCallback
 from optuna.storages._heartbeat import fail_stale_trials
 from optuna.storages._in_memory import InMemoryStorage
 from optuna.storages._rdb.storage import RDBStorage
